% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_testing.R
\name{assoc_testing}
\alias{assoc_testing}
\title{take a distribution of PRS and a Phenotype, the name of the analysis and eventual confounders
return a data frame showing the regression of the PRS on the Phenotype
PRS_snps	Phenotype	Sex_specificity	N_cases	N_controls	OR	lower_CI	upper_CI	P_value}
\usage{
assoc_testing(
  file = NA,
  outcome = NA,
  pheno = pheno,
  prsice = F,
  double_reg = F
)
}
\arguments{
\item{x}{un vecteur numerique}
}
\value{
la fonction renvoie la moyenne d'un vecteur
}
\description{
take a distribution of PRS and a Phenotype, the name of the analysis and eventual confounders
return a data frame showing the regression of the PRS on the Phenotype
PRS_snps	Phenotype	Sex_specificity	N_cases	N_controls	OR	lower_CI	upper_CI	P_value
}
