% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{assoc}
\alias{assoc}
\title{Association of a PRS distribution with a Phenotype}
\usage{
assoc(
  df = NULL,
  prs_col = "SCORESUM",
  phenotype_col = "Phenotype",
  scale = TRUE,
  covar_col = NA,
  log = ""
)
}
\arguments{
\item{df}{a dataframe with individuals on each row, at least one ID column,
one column PRS (continuous variable) and one with phenotype (continuous or discrete)}

\item{prs_col}{a character specifying the PRS column name}

\item{phenotype_col}{a character specifying the Phenotype column name}

\item{scale}{a boolean specifying if scaling of PRS should be done before testing}

\item{covar_col}{a character vector specifying the covariate column names (facultative)}

\item{log}{a connection, or a character string naming the file to print to.
If "" (by default), it prints to the standard output connection, the console unless redirected by sink.}
}
\value{
return a data frame showing the association of the PRS on the Phenotype
with 'PRS','Phenotype','Covar','N_cases','N_controls','N','OR','SE','lower_CI','upper_CI','P_value'
}
\description{
\code{assoc()} take a distribution of PRS, a Phenotype and eventual Confounders
return a data frame showing the association of PRS on the Phenotype
}
