% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decileboxplot.R
\name{decileboxplot}
\alias{decileboxplot}
\title{Deciles BoxPlot from a PRS Association with a Continuous Phenotype}
\usage{
decileboxplot(df = NULL, prs_col = "SCORESUM", phenotype_col = "Phenotype")
}
\arguments{
\item{df}{a dataframe with individuals on each row, at least one column PRS
(continuous variable) and one with phenotype (continuous or categorical)}

\item{prs_col}{a character specifying the PRS column name}

\item{phenotype_col}{a character specifying the Continuous Phenotype column name}
}
\value{
return a figure of results in the format ggplot2 object
}
\description{
\code{decileboxplot()} take a distribution of PRS, a Continuous Phenotype
return a plot (ggplot2 object) with deciles of PRS in x and Boxplot of
the Phenotype in y
}
