% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decileboxplot.R
\name{decileboxplot}
\alias{decileboxplot}
\title{Deciles BoxPlot from a PRS Association}
\usage{
decileboxplot(
  df = NULL,
  prs_col = "SCORESUM",
  phenotype_col = "Phenotype",
  scale = T,
  continuous_metric = NA,
  filename = NA
)
}
\arguments{
\item{df}{a dataframe with individuals on each row, at least one column PRS
(continuous variable) and one with phenotype (continuous or categorical)}

\item{prs_col}{a character specifying the PRS column name}

\item{phenotype_col}{a character specifying the Phenotype column name}

\item{scale}{a boolean specifying if scaling of PRS should be done before plotting}

\item{continuous_metric}{a facultative character specifying what metric to
use for continuous Phenotype, only two options: "median" or "mean"}

\item{filename}{a facultative character, specifying the path and file name
where to store the plot}
}
\value{
return a figure of results in the format ggplot2 object
}
\description{
\code{decileboxplot()} take a distribution of PRS, a Phenotype and eventual Confounders
return a plot (ggplot2 object) with deciles of PRS in x and Prevalence/Median
/Mean of the Phenotype in y
}
