% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityplot.R
\name{densityplot}
\alias{densityplot}
\title{Density Plot from a PRS Association}
\usage{
densityplot(
  df = NULL,
  prs_col = "SCORESUM",
  phenotype_col = "Phenotype",
  scale = T,
  threshold = NA,
  filename = NA
)
}
\arguments{
\item{df}{a dataframe with individuals on each row, at least one column PRS
(continuous variable) and one with phenotype (continuous or categorical)}

\item{prs_col}{a character specifying the PRS column name}

\item{phenotype_col}{a character specifying the Phenotype column name}

\item{scale}{a boolean specifying if scaling of PRS should be done before plotting}

\item{threshold}{a facultative numeric specifying for continuous Phenotype
the Threshold to consider individuals as Cases/Constrols as following:
Phenotype > Threshold = Case
Phenotype < Threshold = Control}

\item{filename}{a facultative character, specifying the path and file name
where to store the plot}
}
\value{
return a figure of results in the format ggplot2 object
}
\description{
\code{densityplot()} take a distribution of PRS, a Phenotype and eventual Confounders
return a plot (ggplot2 object) with density of PRS in x by Categories of the
Phenotype
}
