% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiassoc.R
\name{multiassoc}
\alias{multiassoc}
\title{take a distribution of PRS and multiple Phenotypes, the name of the analysis and eventual confounders
return a data frame showing the regression of the unique PRS on the list of Phenotypes}
\usage{
multiassoc(
  df = NA,
  prs_col = "SCORESUM",
  phenotype_col = "Phenotype",
  scale = TRUE,
  covar_col = NA
)
}
\arguments{
\item{df}{a dataframe with individuals on each row, at least one column PRS
(continuous variable) and one with phenotype (continuous or categorical)}

\item{prs_col}{a character specifying the PRS column name}

\item{phenotype_col}{a character specifying the Phenotype column name}

\item{scale}{a boolean specifying if scaling of PRS should be done before testing}

\item{covar_col}{a character vector specifying the covariate column names (facultative)}
}
\value{
return a data frame showing the regression of the PRS on the Phenotypes
with 'PRS','Phenotype','N_cases','N_controls','N','OR','SE','lower_CI','upper_CI','P_value'
}
\description{
take a distribution of PRS and multiple Phenotypes, the name of the analysis and eventual confounders
return a data frame showing the regression of the unique PRS on the list of Phenotypes
}
