% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiassoc.R
\name{multiassoc}
\alias{multiassoc}
\title{Multiple PRS Associations from a Data Frame}
\usage{
multiassoc(df = NULL, assoc_table = NULL, scale = TRUE, covar_col = NA)
}
\arguments{
\item{df}{a dataframe with individuals on each row, at least one column PRS
(continuous variable) and one with phenotype (continuous or categorical)}

\item{assoc_table}{a dataframe or matrix specifying the associations to
make from df. Each row should have the format 'PRS', 'Phenotype'}

\item{scale}{a boolean specifying if scaling of PRS should be done before testing}

\item{covar_col}{a character vector specifying the covariate column names (facultative)}
}
\value{
return a data frame showing the association of the PRS(s) on the Phenotype(s)
with 'PRS','Phenotype','Covar','N_cases','N_controls','N','OR','SE','lower_CI','upper_CI','P_value'
}
\description{
\code{multiassoc()} take a data frame with distribution(s) of PRS and Phenotype(s),
and a table of associations to make from this data frame
return a data frame showing the association results
}
